requirements.txt
typer
python-gitlab
rich
maya
attrs
cattrs
diskcache
